requirements.txt
fastapi
uvicorn
pytest
httpx
sqlalchemy
psycopg2-binary
pydantic
python-dotenv